+++ frontend/src/components/DefenceBox/DefenceConfigurationRadioButton.tsx
